# contracts/api.yaml
openapi: 3.1.0
info:
  title: Evidence API
  description: Contract-first API for Evidence runtime implementations.
  license:
    name: MIT
    identifier: MIT
  version: 0.1.0
servers:
  - url: http://127.0.0.1:3000
    description: Rust Axum local server
paths:
...
  /api/workspaces/{workspaceId}/diagrams/{diagramId}:
    get:
      tags: []
      operationId: get_diagram
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: diagramId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Diagram resource
          content:
            application/vnd.evidence.diagram+json:
              schema:
                $ref: '#/components/schemas/DiagramResource'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
    put:
      tags: []
      operationId: update_diagram
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: diagramId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDiagramInput'
        required: true
      responses:
        '200':
          description: Updated diagram
          content:
            application/vnd.evidence.diagram+json:
              schema:
                $ref: '#/components/schemas/DiagramResource'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
    delete:
      tags: []
      operationId: delete_diagram
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: diagramId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Diagram delete result
          content:
            application/vnd.evidence.diagram+json:
              schema:
                $ref: '#/components/schemas/DeletedResult'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
  /api/workspaces/{workspaceId}/diagrams/{diagramId}/nodes:
    get:
      tags: []
      operationId: list_diagram_nodes
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: diagramId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Diagram node collection
          content:
            application/vnd.evidence.nodes+json:
              schema:
                $ref: '#/components/schemas/NodeCollectionResource'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
    post:
      tags: []
      operationId: create_diagram_node
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: diagramId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeInput'
        required: true
      responses:
        '200':
          description: Created diagram node
          content:
            application/vnd.evidence.nodes+json:
              schema:
                $ref: '#/components/schemas/NodeResource'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
  /api/workspaces/{workspaceId}/diagrams/{diagramId}/nodes/{nodeId}:
    get:
      tags: []
      operationId: get_diagram_node
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: diagramId
          in: path
          required: true
          schema:
            type: string
        - name: nodeId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Diagram node resource
          content:
            application/vnd.evidence.node+json:
              schema:
                $ref: '#/components/schemas/NodeResource'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
    put:
      tags: []
      operationId: update_diagram_node
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: diagramId
          in: path
          required: true
          schema:
            type: string
        - name: nodeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeInput'
        required: true
      responses:
        '200':
          description: Updated diagram node
          content:
            application/vnd.evidence.node+json:
              schema:
                $ref: '#/components/schemas/NodeResource'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
    delete:
      tags: []
      operationId: delete_diagram_node
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: diagramId
          in: path
          required: true
          schema:
            type: string
        - name: nodeId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Diagram node delete result
          content:
            application/vnd.evidence.node+json:
              schema:
                $ref: '#/components/schemas/DeletedResult'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
  /api/workspaces/{workspaceId}/diagrams/{diagramId}/edges:
    get:
      tags: []
      operationId: list_diagram_edges
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: diagramId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Diagram edge collection
          content:
            application/vnd.evidence.edges+json:
              schema:
                $ref: '#/components/schemas/EdgeCollectionResource'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
    post:
      tags: []
      operationId: create_diagram_edge
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: diagramId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EdgeInput'
        required: true
      responses:
        '200':
          description: Created diagram edge
          content:
            application/vnd.evidence.edges+json:
              schema:
                $ref: '#/components/schemas/EdgeResource'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
...
  /api/workspaces/{workspaceId}/diagrams/{diagramId}/versions:
    get:
      tags: []
      operationId: list_diagram_versions
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: diagramId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Diagram version collection
          content:
            application/vnd.evidence.diagram-versions+json:
              schema:
                $ref: '#/components/schemas/DiagramVersionCollectionResource'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
    post:
      tags: []
      operationId: create_diagram_version
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: diagramId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Created diagram version
          content:
            application/vnd.evidence.diagram-versions+json:
              schema:
                $ref: '#/components/schemas/DiagramVersionResource'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
...
components:
  schemas:
...
    BTreeMap:
      type: object
      additionalProperties:
        type: object
        required:
          - href
        properties:
          href:
            type: string
      propertyNames:
        type: string
...
    DeletedResult:
      type: object
      required:
        - deleted
      properties:
        deleted:
          type: boolean
...
    DiagramResource:
      type: object
      required:
        - _links
        - _templates
        - id
        - title
        - viewport
        - createdAt
        - updatedAt
      properties:
        _links:
          $ref: '#/components/schemas/BTreeMap'
        _templates:
          $ref: '#/components/schemas/BTreeMap'
        createdAt:
          type: string
        id:
          type: string
        title:
          type: string
        updatedAt:
          type: string
        viewport:
          $ref: '#/components/schemas/Viewport'
...
    EdgeCollectionResource:
      type: object
      required:
        - _links
        - _embedded
      properties:
        _embedded:
          $ref: '#/components/schemas/EdgeCollectionEmbedded'
        _links:
          $ref: '#/components/schemas/BTreeMap'
...
    ErrorBody:
      type: object
      required:
        - error
      properties:
        error:
          type: string
...
    NodeCollectionResource:
      type: object
      required:
        - _links
        - _embedded
      properties:
        _embedded:
          $ref: '#/components/schemas/NodeCollectionEmbedded'
        _links:
          $ref: '#/components/schemas/BTreeMap'
...
    UpdateDiagramInput:
      type: object
      properties:
        title:
          type:
            - string
            - 'null'
        viewport:
          oneOf:
            - type: 'null'
            - $ref: '#/components/schemas/Viewport'
        viewport.x:
          type:
            - number
            - 'null'
          format: double
        viewport.y:
          type:
            - number
            - 'null'
          format: double
        viewport.zoom:
          type:
            - number
            - 'null'
          format: double
...